[FLINK-21417][core] Move some access methods from HybridMemorySegment to MemorySegment.
This is in preparation for turning HybridMemorySegment into OffHeapMemorySegment, with all on-heap related logics moved out.

The moved methods all have common implementations that are applicable for both heap and off-heap segments.
- Some of the methods are overridden in HeapMemorySegment for efficiency. However, there's no harm to keep the commonly applicable implementations in the base class, despite they may only be used for the off-heap segments.
- `put(DataInput,int,int)` and `get(DataOutput,int,int)` are partially moved. Only code paths applicable for both heap and off-heap segments are moved.
