+++ flink-core/src/main/java/org/apache/flink/core/memory/HybridMemorySegment.java
@@ -27,10 +27,7 @@ import javax.annotation.Nullable;
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
-import java.nio.BufferOverflowException;
-import java.nio.BufferUnderflowException;
 import java.nio.ByteBuffer;
-import java.nio.ReadOnlyBufferException;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
@@ -166,229 +163,33 @@ public final class HybridMemorySegment extends MemorySegment {
         }
     }
 
-    // ------------------------------------------------------------------------
-    //  Random Access get() and put() methods
-    // ------------------------------------------------------------------------
-
-    @Override
-    public final byte get(int index) {
-        final long pos = address + index;
-        if (index >= 0 && pos < addressLimit) {
-            return UNSAFE.getByte(heapMemory, pos);
-        } else if (address > addressLimit) {
-            throw new IllegalStateException("segment has been freed");
-        } else {
-            // index is in fact invalid
-            throw new IndexOutOfBoundsException();
-        }
-    }
-
-    @Override
-    public final void put(int index, byte b) {
-        final long pos = address + index;
-        if (index >= 0 && pos < addressLimit) {
-            UNSAFE.putByte(heapMemory, pos, b);
-        } else if (address > addressLimit) {
-            throw new IllegalStateException("segment has been freed");
-        } else {
-            // index is in fact invalid
-            throw new IndexOutOfBoundsException();
-        }
-    }
-
-    @Override
-    public final void get(int index, byte[] dst) {
-        get(index, dst, 0, dst.length);
-    }
-
-    @Override
-    public final void put(int index, byte[] src) {
-        put(index, src, 0, src.length);
-    }
-
-    @Override
-    public final void get(int index, byte[] dst, int offset, int length) {
-        // check the byte array offset and length and the status
-        if ((offset | length | (offset + length) | (dst.length - (offset + length))) < 0) {
-            throw new IndexOutOfBoundsException();
-        }
-
-        final long pos = address + index;
-        if (index >= 0 && pos <= addressLimit - length) {
-            final long arrayAddress = BYTE_ARRAY_BASE_OFFSET + offset;
-            UNSAFE.copyMemory(heapMemory, pos, dst, arrayAddress, length);
-        } else if (address > addressLimit) {
-            throw new IllegalStateException("segment has been freed");
-        } else {
-            throw new IndexOutOfBoundsException(
-                    String.format(
-                            "pos: %d, length: %d, index: %d, offset: %d",
-                            pos, length, index, offset));
-        }
-    }
-
-    @Override
-    public final void put(int index, byte[] src, int offset, int length) {
-        // check the byte array offset and length
-        if ((offset | length | (offset + length) | (src.length - (offset + length))) < 0) {
-            throw new IndexOutOfBoundsException();
-        }
-
-        final long pos = address + index;
-
-        if (index >= 0 && pos <= addressLimit - length) {
-            final long arrayAddress = BYTE_ARRAY_BASE_OFFSET + offset;
-            UNSAFE.copyMemory(src, arrayAddress, heapMemory, pos, length);
-        } else if (address > addressLimit) {
-            throw new IllegalStateException("segment has been freed");
-        } else {
-            // index is in fact invalid
-            throw new IndexOutOfBoundsException();
-        }
-    }
-
-    @Override
-    public final boolean getBoolean(int index) {
-        return get(index) != 0;
-    }
-
-    @Override
-    public final void putBoolean(int index, boolean value) {
-        put(index, (byte) (value ? 1 : 0));
-    }
-
     // -------------------------------------------------------------------------
     //  Bulk Read and Write Methods
     // -------------------------------------------------------------------------
 
     @Override
     public final void get(DataOutput out, int offset, int length) throws IOException {
-        if (address <= addressLimit) {
-            if (heapMemory != null) {
+        if (heapMemory != null) {
+            if (address <= addressLimit) {
                 out.write(heapMemory, offset, length);
             } else {
-                while (length >= 8) {
-                    out.writeLong(getLongBigEndian(offset));
-                    offset += 8;
-                    length -= 8;
-                }
-
-                while (length > 0) {
-                    out.writeByte(get(offset));
-                    offset++;
-                    length--;
-                }
+                throw new IllegalStateException("segment has been freed");
             }
         } else {
-            throw new IllegalStateException("segment has been freed");
+            super.get(out, offset, length);
         }
     }
 
     @Override
     public final void put(DataInput in, int offset, int length) throws IOException {
-        if (address <= addressLimit) {
-            if (heapMemory != null) {
+        if (heapMemory != null) {
+            if (address <= addressLimit) {
                 in.readFully(heapMemory, offset, length);
             } else {
-                while (length >= 8) {
-                    putLongBigEndian(offset, in.readLong());
-                    offset += 8;
-                    length -= 8;
-                }
-                while (length > 0) {
-                    put(offset, in.readByte());
-                    offset++;
-                    length--;
-                }
-            }
-        } else {
-            throw new IllegalStateException("segment has been freed");
-        }
-    }
-
-    @Override
-    public final void get(int offset, ByteBuffer target, int numBytes) {
-        // check the byte array offset and length
-        if ((offset | numBytes | (offset + numBytes)) < 0) {
-            throw new IndexOutOfBoundsException();
-        }
-        if (target.isReadOnly()) {
-            throw new ReadOnlyBufferException();
-        }
-
-        final int targetOffset = target.position();
-        final int remaining = target.remaining();
-
-        if (remaining < numBytes) {
-            throw new BufferOverflowException();
-        }
-
-        if (target.isDirect()) {
-            // copy to the target memory directly
-            final long targetPointer = getByteBufferAddress(target) + targetOffset;
-            final long sourcePointer = address + offset;
-
-            if (sourcePointer <= addressLimit - numBytes) {
-                UNSAFE.copyMemory(heapMemory, sourcePointer, null, targetPointer, numBytes);
-                target.position(targetOffset + numBytes);
-            } else if (address > addressLimit) {
                 throw new IllegalStateException("segment has been freed");
-            } else {
-                throw new IndexOutOfBoundsException();
             }
-        } else if (target.hasArray()) {
-            // move directly into the byte array
-            get(offset, target.array(), targetOffset + target.arrayOffset(), numBytes);
-
-            // this must be after the get() call to ensue that the byte buffer is not
-            // modified in case the call fails
-            target.position(targetOffset + numBytes);
         } else {
-            // other types of byte buffers
-            throw new IllegalArgumentException(
-                    "The target buffer is not direct, and has no array.");
-        }
-    }
-
-    @Override
-    public final void put(int offset, ByteBuffer source, int numBytes) {
-        // check the byte array offset and length
-        if ((offset | numBytes | (offset + numBytes)) < 0) {
-            throw new IndexOutOfBoundsException();
-        }
-
-        final int sourceOffset = source.position();
-        final int remaining = source.remaining();
-
-        if (remaining < numBytes) {
-            throw new BufferUnderflowException();
-        }
-
-        if (source.isDirect()) {
-            // copy to the target memory directly
-            final long sourcePointer = getByteBufferAddress(source) + sourceOffset;
-            final long targetPointer = address + offset;
-
-            if (targetPointer <= addressLimit - numBytes) {
-                UNSAFE.copyMemory(null, sourcePointer, heapMemory, targetPointer, numBytes);
-                source.position(sourceOffset + numBytes);
-            } else if (address > addressLimit) {
-                throw new IllegalStateException("segment has been freed");
-            } else {
-                throw new IndexOutOfBoundsException();
-            }
-        } else if (source.hasArray()) {
-            // move directly into the byte array
-            put(offset, source.array(), sourceOffset + source.arrayOffset(), numBytes);
-
-            // this must be after the get() call to ensue that the byte buffer is not
-            // modified in case the call fails
-            source.position(sourceOffset + numBytes);
-        } else {
-            // other types of byte buffers
-            for (int i = 0; i < numBytes; i++) {
-                put(offset++, source.get());
-            }
+            super.put(in, offset, length);
         }
     }
