+++ flink-core/src/main/java/org/apache/flink/core/memory/MemorySegment.java
@@ -23,11 +23,15 @@ import org.apache.flink.annotation.Internal;
 import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
+import java.nio.BufferOverflowException;
+import java.nio.BufferUnderflowException;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import java.nio.ReadOnlyBufferException;
 import java.util.function.Consumer;
 import java.util.function.Function;
+
+import static org.apache.flink.core.memory.MemoryUtils.getByteBufferAddress;
 
 /**
  * This class represents a piece of memory managed by Flink. The segment may be backed by heap
@@ -100,11 +104,11 @@ public abstract class MemorySegment {
 
     /** The unsafe handle for transparent memory copied (heap / off-heap). */
     @SuppressWarnings("restriction")
-    protected static final sun.misc.Unsafe UNSAFE = MemoryUtils.UNSAFE;
+    private static final sun.misc.Unsafe UNSAFE = MemoryUtils.UNSAFE;
 
     /** The beginning of the byte array contents, relative to the byte array object. */
     @SuppressWarnings("restriction")
-    protected static final long BYTE_ARRAY_BASE_OFFSET = UNSAFE.arrayBaseOffset(byte[].class);
+    private static final long BYTE_ARRAY_BASE_OFFSET = UNSAFE.arrayBaseOffset(byte[].class);
 
     /**
      * Constant that flags the byte order. Because this is a boolean constant, the JIT compiler can
@@ -309,7 +313,17 @@ public abstract class MemorySegment {
      * @throws IndexOutOfBoundsException Thrown, if the index is negative, or larger or equal to the
      *     size of the memory segment.
      */
-    public abstract byte get(int index);
+    public byte get(int index) {
+        final long pos = address + index;
+        if (index >= 0 && pos < addressLimit) {
+            return UNSAFE.getByte(heapMemory, pos);
+        } else if (address > addressLimit) {
+            throw new IllegalStateException("segment has been freed");
+        } else {
+            // index is in fact invalid
+            throw new IndexOutOfBoundsException();
+        }
+    }
 
     /**
      * Writes the given byte into this buffer at the given position.
@@ -319,7 +333,17 @@ public abstract class MemorySegment {
      * @throws IndexOutOfBoundsException Thrown, if the index is negative, or larger or equal to the
      *     size of the memory segment.
      */
-    public abstract void put(int index, byte b);
+    public void put(int index, byte b) {
+        final long pos = address + index;
+        if (index >= 0 && pos < addressLimit) {
+            UNSAFE.putByte(heapMemory, pos, b);
+        } else if (address > addressLimit) {
+            throw new IllegalStateException("segment has been freed");
+        } else {
+            // index is in fact invalid
+            throw new IndexOutOfBoundsException();
+        }
+    }
 
     /**
      * Bulk get method. Copies dst.length memory from the specified position to the destination
@@ -331,7 +355,9 @@ public abstract class MemorySegment {
      *     data between the index and the memory segment end is not enough to fill the destination
      *     array.
      */
-    public abstract void get(int index, byte[] dst);
+    public final void get(int index, byte[] dst) {
+        get(index, dst, 0, dst.length);
+    }
 
     /**
      * Bulk put method. Copies src.length memory from the source memory into the memory segment
@@ -343,7 +369,9 @@ public abstract class MemorySegment {
      *     the array size exceed the amount of memory between the index and the memory segment's
      *     end.
      */
-    public abstract void put(int index, byte[] src);
+    public final void put(int index, byte[] src) {
+        put(index, src, 0, src.length);
+    }
 
     /**
      * Bulk get method. Copies length memory from the specified position to the destination memory,
@@ -357,7 +385,25 @@ public abstract class MemorySegment {
      *     requested number of bytes exceed the amount of memory between the index and the memory
      *     segment's end.
      */
-    public abstract void get(int index, byte[] dst, int offset, int length);
+    public void get(int index, byte[] dst, int offset, int length) {
+        // check the byte array offset and length and the status
+        if ((offset | length | (offset + length) | (dst.length - (offset + length))) < 0) {
+            throw new IndexOutOfBoundsException();
+        }
+
+        final long pos = address + index;
+        if (index >= 0 && pos <= addressLimit - length) {
+            final long arrayAddress = BYTE_ARRAY_BASE_OFFSET + offset;
+            UNSAFE.copyMemory(heapMemory, pos, dst, arrayAddress, length);
+        } else if (address > addressLimit) {
+            throw new IllegalStateException("segment has been freed");
+        } else {
+            throw new IndexOutOfBoundsException(
+                    String.format(
+                            "pos: %d, length: %d, index: %d, offset: %d",
+                            pos, length, index, offset));
+        }
+    }
 
     /**
      * Bulk put method. Copies length memory starting at position offset from the source memory into
@@ -371,7 +417,24 @@ public abstract class MemorySegment {
      *     the array portion to copy exceed the amount of memory between the index and the memory
      *     segment's end.
      */
-    public abstract void put(int index, byte[] src, int offset, int length);
+    public void put(int index, byte[] src, int offset, int length) {
+        // check the byte array offset and length
+        if ((offset | length | (offset + length) | (src.length - (offset + length))) < 0) {
+            throw new IndexOutOfBoundsException();
+        }
+
+        final long pos = address + index;
+
+        if (index >= 0 && pos <= addressLimit - length) {
+            final long arrayAddress = BYTE_ARRAY_BASE_OFFSET + offset;
+            UNSAFE.copyMemory(src, arrayAddress, heapMemory, pos, length);
+        } else if (address > addressLimit) {
+            throw new IllegalStateException("segment has been freed");
+        } else {
+            // index is in fact invalid
+            throw new IndexOutOfBoundsException();
+        }
+    }
 
     /**
      * Reads one byte at the given position and returns its boolean representation.
@@ -381,7 +444,9 @@ public abstract class MemorySegment {
      * @throws IndexOutOfBoundsException Thrown, if the index is negative, or larger than the
      *     segment size minus 1.
      */
-    public abstract boolean getBoolean(int index);
+    public final boolean getBoolean(int index) {
+        return get(index) != 0;
+    }
 
     /**
      * Writes one byte containing the byte value into this buffer at the given position.
@@ -391,7 +456,9 @@ public abstract class MemorySegment {
      * @throws IndexOutOfBoundsException Thrown, if the index is negative, or larger than the
      *     segment size minus 1.
      */
-    public abstract void putBoolean(int index, boolean value);
+    public final void putBoolean(int index, boolean value) {
+        put(index, (byte) (value ? 1 : 0));
+    }
 
     /**
      * Reads a char value from the given position, in the system's native byte order.
@@ -1121,7 +1188,23 @@ public abstract class MemorySegment {
     //                     Bulk Read and Write Methods
     // -------------------------------------------------------------------------
 
-    public abstract void get(DataOutput out, int offset, int length) throws IOException;
+    public void get(DataOutput out, int offset, int length) throws IOException {
+        if (address <= addressLimit) {
+            while (length >= 8) {
+                out.writeLong(getLongBigEndian(offset));
+                offset += 8;
+                length -= 8;
+            }
+
+            while (length > 0) {
+                out.writeByte(get(offset));
+                offset++;
+                length--;
+            }
+        } else {
+            throw new IllegalStateException("segment has been freed");
+        }
+    }
 
     /**
      * Bulk put method. Copies length memory from the given DataInput to the memory starting at
@@ -1133,7 +1216,22 @@ public abstract class MemorySegment {
      * @throws IOException Thrown, if the DataInput encountered a problem upon reading, such as an
      *     End-Of-File.
      */
-    public abstract void put(DataInput in, int offset, int length) throws IOException;
+    public void put(DataInput in, int offset, int length) throws IOException {
+        if (address <= addressLimit) {
+            while (length >= 8) {
+                putLongBigEndian(offset, in.readLong());
+                offset += 8;
+                length -= 8;
+            }
+            while (length > 0) {
+                put(offset, in.readByte());
+                offset++;
+                length--;
+            }
+        } else {
+            throw new IllegalStateException("segment has been freed");
+        }
+    }
 
     /**
      * Bulk get method. Copies {@code numBytes} bytes from this memory segment, starting at position
@@ -1151,7 +1249,48 @@ public abstract class MemorySegment {
      *     enough space for the bytes.
      * @throws ReadOnlyBufferException If the target buffer is read-only.
      */
-    public abstract void get(int offset, ByteBuffer target, int numBytes);
+    public void get(int offset, ByteBuffer target, int numBytes) {
+        // check the byte array offset and length
+        if ((offset | numBytes | (offset + numBytes)) < 0) {
+            throw new IndexOutOfBoundsException();
+        }
+        if (target.isReadOnly()) {
+            throw new ReadOnlyBufferException();
+        }
+
+        final int targetOffset = target.position();
+        final int remaining = target.remaining();
+
+        if (remaining < numBytes) {
+            throw new BufferOverflowException();
+        }
+
+        if (target.isDirect()) {
+            // copy to the target memory directly
+            final long targetPointer = getByteBufferAddress(target) + targetOffset;
+            final long sourcePointer = address + offset;
+
+            if (sourcePointer <= addressLimit - numBytes) {
+                UNSAFE.copyMemory(heapMemory, sourcePointer, null, targetPointer, numBytes);
+                target.position(targetOffset + numBytes);
+            } else if (address > addressLimit) {
+                throw new IllegalStateException("segment has been freed");
+            } else {
+                throw new IndexOutOfBoundsException();
+            }
+        } else if (target.hasArray()) {
+            // move directly into the byte array
+            get(offset, target.array(), targetOffset + target.arrayOffset(), numBytes);
+
+            // this must be after the get() call to ensue that the byte buffer is not
+            // modified in case the call fails
+            target.position(targetOffset + numBytes);
+        } else {
+            // other types of byte buffers
+            throw new IllegalArgumentException(
+                    "The target buffer is not direct, and has no array.");
+        }
+    }
 
     /**
      * Bulk put method. Copies {@code numBytes} bytes from the given {@code ByteBuffer}, into this
@@ -1169,7 +1308,46 @@ public abstract class MemorySegment {
      *     contain the given number of bytes, or this segment does not have enough space for the
      *     bytes (counting from offset).
      */
-    public abstract void put(int offset, ByteBuffer source, int numBytes);
+    public void put(int offset, ByteBuffer source, int numBytes) {
+        // check the byte array offset and length
+        if ((offset | numBytes | (offset + numBytes)) < 0) {
+            throw new IndexOutOfBoundsException();
+        }
+
+        final int sourceOffset = source.position();
+        final int remaining = source.remaining();
+
+        if (remaining < numBytes) {
+            throw new BufferUnderflowException();
+        }
+
+        if (source.isDirect()) {
+            // copy to the target memory directly
+            final long sourcePointer = getByteBufferAddress(source) + sourceOffset;
+            final long targetPointer = address + offset;
+
+            if (targetPointer <= addressLimit - numBytes) {
+                UNSAFE.copyMemory(null, sourcePointer, heapMemory, targetPointer, numBytes);
+                source.position(sourceOffset + numBytes);
+            } else if (address > addressLimit) {
+                throw new IllegalStateException("segment has been freed");
+            } else {
+                throw new IndexOutOfBoundsException();
+            }
+        } else if (source.hasArray()) {
+            // move directly into the byte array
+            put(offset, source.array(), sourceOffset + source.arrayOffset(), numBytes);
+
+            // this must be after the get() call to ensue that the byte buffer is not
+            // modified in case the call fails
+            source.position(sourceOffset + numBytes);
+        } else {
+            // other types of byte buffers
+            for (int i = 0; i < numBytes; i++) {
+                put(offset++, source.get());
+            }
+        }
+    }
 
     /**
      * Bulk copy method. Copies {@code numBytes} bytes from this memory segment, starting at
